Organize switch related files
