+++ switch/__init__.py
@@ -1,0 +1,9 @@
+from .abstract_switch import AbstractSwitch
+from .sp8t_switch import SP8T_Switch
+from .multi_switch import MultiSwitch
+
+__all__ = [
+    "AbstractSwitch",
+    "SP8T_Switch",
+    "MultiSwitch",
+]
